+++ mux/script/mux_main.cpp
@@ -96,8 +96,8 @@ public:
     virtual MUX_RESULT SumPlayerCommandCount(dbref target, int *pCount);
     virtual MUX_RESULT FetchHeight(dbref target, int *pHeight);
     virtual MUX_RESULT FetchWidth(dbref target, int *pWidth);
-    virtual MUX_RESULT FetchIdle(dbref target, int *pIdle);
-    virtual MUX_RESULT FetchConnect(dbref target, int *pConnect);
+    virtual MUX_RESULT FetchIdle(dbref target, int64_t *pIdle);
+    virtual MUX_RESULT FetchConnect(dbref target, int64_t *pConnect);
 
     // Queries by opaque DESC handle
     virtual MUX_RESULT FindDescBySocket(SOCKET s, DESC **ppDesc);
@@ -306,7 +306,7 @@ MUX_RESULT CScriptConnectionManager::FetchWidth(dbref target, int *pWidth)
     return MUX_S_OK;
 }
 
-MUX_RESULT CScriptConnectionManager::FetchIdle(dbref target, int *pIdle)
+MUX_RESULT CScriptConnectionManager::FetchIdle(dbref target, int64_t *pIdle)
 {
     UNUSED_PARAMETER(target);
     if (nullptr == pIdle) return MUX_E_INVALIDARG;
@@ -314,7 +314,7 @@ MUX_RESULT CScriptConnectionManager::FetchIdle(dbref target, int *pIdle)
     return MUX_S_OK;
 }
 
-MUX_RESULT CScriptConnectionManager::FetchConnect(dbref target, int *pConnect)
+MUX_RESULT CScriptConnectionManager::FetchConnect(dbref target, int64_t *pConnect)
 {
     UNUSED_PARAMETER(target);
     if (nullptr == pConnect) return MUX_E_INVALIDARG;
